Port to VTK version 6.0
git-svn-id: svn+ssh://edison.nersc.gov/project/projectdirs/visit/svn/visit/trunk@20931 18c085ea-50e0-402c-830e-de6fd14e8384

+++ src/avt/Expressions/General/avtPythonExpression.C
@@ -306,7 +306,7 @@ avtPythonExpression::ProcessArguments(ArgsExpr *args,
 
     // get the argument list and # of arguments
     std::vector<ArgExpr*> *arguments = args->GetArgs();
-    int nargs = arguments->size();
+    size_t nargs = arguments->size();
 
     // check for call with improper # of arguments
     if (nargs < 2)
@@ -351,7 +351,7 @@ avtPythonExpression::ProcessArguments(ArgsExpr *args,
                      "from python filter.");
 
     // process the remaining arguments
-    for(int i=0;i< nargs-1;i++)
+    for(size_t i=0;i< nargs-1;i++)
     {
         ExprParseTreeNode *node =  ((*arguments)[i])->GetExpr();
         std::string node_type = node->GetTypeName();
